worker_requirement.txt
celery
ujson
GPUtil
pillow
redis
pyyaml
opencv_python
torch
segment_anything
torchvision
git+https://github.com/IDEA-Research/GroundingDINO
gradio
pandas
scipy
pytorch_lightning==1.7.6
einops
torchmetrics==0.7.3
omegaconf
gitpython
piexif
scikit-image
open_clip_torch
tomesd
psutil
xformers
lark
accelerate
basicsr
blendmodes
clean-fid
gfpgan
inflection
jsonmerge
kornia
numpy
omegaconf
realesrgan
requests
resize-right
safetensors
scikit-image>=0.19
timm
torchdiffeq
torchsde
transformers==4.25.1
clip
opennsfw2
tensorflow
typing_extensions
gradio==3.31.0
fastapi==0.94.0
ultralytics